backend/requirements.txt
# Alternative with latest versions
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# AI/LLM Libraries
google-generativeai==0.3.2
openai==1.10.0
anthropic==0.19.1

# Supabase - let pip resolve the best version
supabase
httpx  # Don't specify version, let pip resolve

# Authentication
PyJWT==2.8.0

# Other utilities
requests==2.31.0
aiohttp==3.9.1